+++ src/metatrain/experimental/space/tests/__init__.py
@@ -8,7 +8,7 @@ DATASET_PATH_PERIODIC = str(
     Path(__file__).parents[5] / "tests/resources/carbon_reduced_100.xyz"
 )
 
-DEFAULT_HYPERS = get_default_hypers("experimental.phace")
+DEFAULT_HYPERS = get_default_hypers("experimental.space")
 MODEL_HYPERS = DEFAULT_HYPERS["model"]
 SPHERICAL_DISK_DATASET_PATH = str(
     Path(__file__).parents[5] / "tests/resources/spherical_disk_dataset.zip"